Struct cbqn::BQNValue

pub struct BQNValue { /* private fields */ }

Represents a BQN value

Implementations

impl BQNValue

pub fn null() -> BQNValue

Constructs a BQN null value @

Examples

BQN!('a', "-", BQNValue::null());

pub fn has_field(&self, field: &str) -> bool

Returns a boolean value indicating whether field exists in a BQN namespace As CBQN requires the searched field name to be in a string with all lowercase letters, this function returns false if it is supplied with a field string that contains uppercase characters.

Panics

If self isn’t a namespace.

pub fn get_field(&self, field: &str) -> Option<BQNValue>

Returns field from a BQN namespace as BQNValue. Returns None if the field cannot be found. As CBQN requires the searched field name to be in a string with all lowercase letters, this function returns None if it is supplied with a field string that contains uppercase characters.

Panics

If self isn’t a namespace.

pub fn call1(&self, x: &BQNValue) -> BQNValue

Calls BQNValue as a function with one argument

pub fn call2(&self, w: &BQNValue, x: &BQNValue) -> BQNValue

Calls BQNValue as a function with two arguments

pub fn bqn_type(&self) -> BQNType

Returns the BQN type of the BQNValue

pub fn to_f64(&self) -> f64

Converts BQNValue into f64

Panics

If self isn’t a number

pub fn to_char(&self) -> Option<char>

Converts BQNValue into char

Returns None if the value is not an Unicode scalar value. Rust chars cannot represent characters that are not Unicode scalar values.

Panics

If self isn’t a BQN character

pub fn to_u32(&self) -> u32

Converts BQNValue into u32

BQN characters can contain values that aren’t Unicode scalar values. Those characters can be converted into a Rust type u32 using this function.

Panics

If self isn’t a BQN character

pub fn to_f64_vec(&self) -> Vec<f64>

Converts BQNValue into a vector of f64s

Panics

If self isn’t a BQN array containing numbers

pub fn to_bqnvalue_vec(&self) -> Vec<BQNValue>

Converts BQNValue into a vector of BQNValues

Panics

If self isn’t a BQN array that contains BQN objects

pub fn to_char_vec(&self) -> Vec<char>

Converts BQNValue into vector of chars

Panics

If self isn’t a BQN array that contains characters

pub fn to_string(&self) -> String

Converts BQNValue into a String

Panics

If self isn’t a BQN array that contains characters

pub fn fn1(func: fn(_: &BQNValue) -> BQNValue) -> BQNValue

Generates a BQNValue from a Rust function

The function receives one argument.

Examples

let add_three = BQNValue::fn1(|x| BQNValue::from(x.to_f64() + 3.0));
assert_eq!(BQN!(3, "{𝕏𝕨}", add_three).to_f64(), 6.0);

Panics

If called from a BQNValue::fn1 or BQNValue::fn2 function

Implementation note

Calling this function will allocate memory that will last for the lifetime of the program. Calling it with two identical closures, but with different lifetimes, will allocate the memory multiple times.

pub fn fn2(func: fn(_: &BQNValue, _: &BQNValue) -> BQNValue) -> BQNValue

Generates a BQNValue from a Rust function

The function receives two arguments.

Examples

let multiply = BQNValue::fn2(|w, x| BQNValue::from(w.to_f64() * x.to_f64()));
assert_eq!(BQN!(multiply, "{𝕎´𝕩}", [1,2,3,4,5]).to_f64(), 120.0);

Panics

If called from a BQNValue::fn1 or BQNValue::fn2 function

Implementation note

Calling this function will allocate memory that will last for the lifetime of the program. Calling it with two identical closures, but with different lifetimes, will allocate the memory multiple times.